Refactor module lifecycle package structure and replace `module` with specialized subpackages (`discovery`, `install`, `scaffold`, `publish`) across the codebase.
