+++ pkg/module/install/composition_test.go
@@ -1,4 +1,4 @@
-package module
+package install
 
 import (
 	"go/ast"
