+++ pkg/module/install/go_runner.go
@@ -1,4 +1,4 @@
-package module
+package install
 
 import (
 	"bytes"
