+++ pkg/module/install/reference.go
@@ -1,4 +1,4 @@
-package module
+package install
 
 import (
 	"fmt"
