+++ pkg/module/install/reference_test.go
@@ -1,4 +1,4 @@
-package module
+package install
 
 import "testing"
