+++ pkg/module/install/scripted_go_runner_test.go
@@ -1,4 +1,4 @@
-package module
+package install
 
 import (
 	"context"
