+++ pkg/module/install/transaction_test.go
@@ -1,4 +1,4 @@
-package module
+package install
 
 import (
 	"errors"
