+++ pkg/module/publish/types.go
@@ -1,0 +1,7 @@
+package publish
+
+// Options controls local publication preparation.
+type Options struct {
+	Directory string
+	Tag       string
+}
