+++ pkg/module/service_test.go
@@ -2,215 +2,112 @@ package module
 
 import (
 	"context"
-	"errors"
-	"reflect"
+	"strings"
 	"testing"
 
-	barnregistry "github.com/MontFerret/barn/pkg/registry"
+	barnpublish "github.com/MontFerret/barn/pkg/publish"
+
+	"github.com/MontFerret/cli/v2/pkg/module/discovery"
+	"github.com/MontFerret/cli/v2/pkg/module/install"
+	modulepublish "github.com/MontFerret/cli/v2/pkg/module/publish"
+	"github.com/MontFerret/cli/v2/pkg/module/scaffold"
 )
 
-func TestServiceSearchUsesBarnFilteringAndProjectsDetails(t *testing.T) {
-	registry := &fakeRegistry{
-		summaries: []barnregistry.ModuleSummary{
-			{ID: "zeta/http", Latest: "1.0.0"},
-			{ID: "acme/sqlite"},
-		},
-		modules: map[string]*barnregistry.Module{
-			"zeta/http": {
-				ID: "zeta/http", Description: "HTTP tools",
-				Versions: []barnregistry.VersionSummary{{Version: "1.0.0"}},
-			},
-			"acme/sqlite": {
-				ID: "acme/sqlite", Description: "Embedded database",
-				Versions: []barnregistry.VersionSummary{{Version: "2.0.0-rc.1"}},
-			},
-		},
-	}
-	service := NewService(registry, nil, nil)
+type fakeLifecycle struct {
+	query          string
+	name           string
+	installOptions install.Options
+	createOptions  scaffold.Options
+	publishOptions modulepublish.Options
+	searchResults  []discovery.SearchResult
+	info           *discovery.ModuleInfo
+	installResult  *install.Result
+	createResult   *scaffold.Result
+	publishResult  *barnpublish.Result
+}
 
-	results, err := service.Search(context.Background(), "SQLITE")
-	if err != nil {
-		t.Fatal(err)
+func (f *fakeLifecycle) Search(_ context.Context, query string) ([]discovery.SearchResult, error) {
+	f.query = query
+	return f.searchResults, nil
+}
+
+func (f *fakeLifecycle) Info(_ context.Context, name string) (*discovery.ModuleInfo, error) {
+	f.name = name
+	return f.info, nil
+}
+
+func (f *fakeLifecycle) Install(_ context.Context, options install.Options) (*install.Result, error) {
+	f.installOptions = options
+	return f.installResult, nil
+}
+
+func (f *fakeLifecycle) Create(_ context.Context, options scaffold.Options) (*scaffold.Result, error) {
+	f.createOptions = options
+	return f.createResult, nil
+}
+
+func (f *fakeLifecycle) Prepare(_ context.Context, options modulepublish.Options) (*barnpublish.Result, error) {
+	f.publishOptions = options
+	return f.publishResult, nil
+}
+
+func TestServiceDelegatesToWorkflowComponents(t *testing.T) {
+	lifecycle := &fakeLifecycle{
+		searchResults: []discovery.SearchResult{{Name: "acme/widget"}},
+		info:          &discovery.ModuleInfo{Name: "acme/widget"},
+		installResult: &install.Result{ID: "acme/widget"},
+		createResult:  &scaffold.Result{Directory: "widget"},
+		publishResult: &barnpublish.Result{Kind: barnpublish.NewModule},
+	}
+	service := NewService(lifecycle, lifecycle, lifecycle, lifecycle)
+	ctx := context.Background()
+
+	searchResults, err := service.Search(ctx, "widget")
+	if err != nil || len(searchResults) != 1 || searchResults[0].Name != "acme/widget" || lifecycle.query != "widget" {
+		t.Fatalf("unexpected search delegation: results=%#v query=%q err=%v", searchResults, lifecycle.query, err)
 	}
 
-	if registry.searchOptions != (barnregistry.SearchOptions{Query: "SQLITE"}) {
-		t.Fatalf("unexpected search options: %#v", registry.searchOptions)
+	info, err := service.Info(ctx, "acme/widget")
+	if err != nil || info != lifecycle.info || lifecycle.name != "acme/widget" {
+		t.Fatalf("unexpected info delegation: info=%#v name=%q err=%v", info, lifecycle.name, err)
 	}
 
-	want := []SearchResult{
-		{Name: "acme/sqlite", Version: "2.0.0-rc.1", Description: "Embedded database"},
-		{Name: "zeta/http", Version: "1.0.0", Description: "HTTP tools"},
+	installOptions := install.Options{Reference: "acme/widget@1.0.0", Directory: "."}
+	installed, err := service.Install(ctx, installOptions)
+	if err != nil || installed != lifecycle.installResult || lifecycle.installOptions != installOptions {
+		t.Fatalf("unexpected install delegation: result=%#v options=%#v err=%v", installed, lifecycle.installOptions, err)
 	}
-	if !reflect.DeepEqual(results, want) {
-		t.Fatalf("unexpected results: %#v", results)
+
+	createOptions := scaffold.Options{Name: "acme/widget", GoModule: "example.com/widget"}
+	created, err := service.Create(ctx, createOptions)
+	if err != nil || created != lifecycle.createResult || lifecycle.createOptions != createOptions {
+		t.Fatalf("unexpected create delegation: result=%#v options=%#v err=%v", created, lifecycle.createOptions, err)
+	}
+
+	publishOptions := modulepublish.Options{Directory: ".", Tag: "v1.0.0"}
+	published, err := service.Publish(ctx, publishOptions)
+	if err != nil || published != lifecycle.publishResult || lifecycle.publishOptions != publishOptions {
+		t.Fatalf("unexpected publish delegation: result=%#v options=%#v err=%v", published, lifecycle.publishOptions, err)
 	}
 }
 
-func TestServiceSearchForwardsDescriptionQuery(t *testing.T) {
-	registry := &fakeRegistry{
-		search: func(_ context.Context, options barnregistry.SearchOptions) ([]barnregistry.ModuleSummary, error) {
-			if options != (barnregistry.SearchOptions{Query: "ai"}) {
-				t.Fatalf("unexpected search options: %#v", options)
-			}
+func TestServiceRejectsUnconfiguredComponents(t *testing.T) {
+	service := NewService(nil, nil, nil, nil)
+	ctx := context.Background()
 
-			return []barnregistry.ModuleSummary{{ID: "montferret/llm"}}, nil
-		},
-		modules: map[string]*barnregistry.Module{
-			"montferret/llm": {
-				ID:          "montferret/llm",
-				Description: "Text generation under AI::LLM for Ferret.",
-				Versions:    []barnregistry.VersionSummary{{Version: "1.0.0-rc.4"}},
-			},
-		},
+	if _, err := service.Search(ctx, ""); err == nil || !strings.Contains(err.Error(), "discovery") {
+		t.Fatalf("expected discovery configuration error, got %v", err)
 	}
-	service := NewService(registry, nil, nil)
-
-	results, err := service.Search(context.Background(), "ai")
-	if err != nil {
-		t.Fatal(err)
+	if _, err := service.Info(ctx, "acme/widget"); err == nil || !strings.Contains(err.Error(), "discovery") {
+		t.Fatalf("expected discovery configuration error, got %v", err)
 	}
-
-	want := []SearchResult{{
-		Name: "montferret/llm", Version: "1.0.0-rc.4", Description: "Text generation under AI::LLM for Ferret.",
-	}}
-	if !reflect.DeepEqual(results, want) {
-		t.Fatalf("unexpected results: %#v", results)
+	if _, err := service.Install(ctx, install.Options{}); err == nil || !strings.Contains(err.Error(), "installer") {
+		t.Fatalf("expected installer configuration error, got %v", err)
 	}
-}
-
-func TestServiceSearchPropagatesBarnAndCancellationErrors(t *testing.T) {
-	t.Run("typed Barn error", func(t *testing.T) {
-		failure := &barnregistry.HTTPError{URL: "https://registry.example/index.json", StatusCode: 503, Status: "503 Service Unavailable"}
-		service := NewService(&fakeRegistry{searchErr: failure}, nil, nil)
-
-		_, err := service.Search(context.Background(), "")
-		var httpError *barnregistry.HTTPError
-		if !errors.As(err, &httpError) || httpError != failure {
-			t.Fatalf("expected Barn HTTP error, got %v", err)
-		}
-	})
-
-	t.Run("cancellation", func(t *testing.T) {
-		service := NewService(&fakeRegistry{search: func(ctx context.Context, _ barnregistry.SearchOptions) ([]barnregistry.ModuleSummary, error) {
-			<-ctx.Done()
-
-			return nil, ctx.Err()
-		}}, nil, nil)
-		ctx, cancel := context.WithCancel(context.Background())
-		cancel()
-
-		_, err := service.Search(ctx, "")
-		if !errors.Is(err, context.Canceled) {
-			t.Fatalf("expected context cancellation, got %v", err)
-		}
-	})
-}
-
-func TestServiceInfoSelectsLatestOrNewest(t *testing.T) {
-	registry := &fakeRegistry{
-		modules: map[string]*barnregistry.Module{
-			"acme/sqlite": {
-				ID:          "acme/sqlite",
-				Description: "SQLite",
-				Latest:      "1.0.0",
-				Versions: []barnregistry.VersionSummary{
-					{Version: "2.0.0-rc.1"},
-					{Version: "1.0.0"},
-				},
-			},
-			"acme/preview": {
-				ID:          "acme/preview",
-				Description: "Preview",
-				Versions:    []barnregistry.VersionSummary{{Version: "1.0.0-rc.1"}},
-			},
-		},
-		versions: map[string]*barnregistry.Version{
-			"acme/sqlite@1.0.0": {
-				ID: "acme/sqlite", Version: "1.0.0", Namespace: "DB::SQLITE", Ferret: ">=2.0.0 <3.0.0",
-				Source:  barnregistry.Source{Repository: "https://example.com/acme/sqlite", Path: "modules/sqlite", Commit: "abc"},
-				Content: map[string]string{"documentation": "https://registry.example/docs.md"},
-			},
-			"acme/preview@1.0.0-rc.1": {
-				ID: "acme/preview", Version: "1.0.0-rc.1", Namespace: "PREVIEW",
-				Source: barnregistry.Source{Repository: "https://example.com/acme/preview", Commit: "def"},
-			},
-		},
+	if _, err := service.Create(ctx, scaffold.Options{}); err == nil || !strings.Contains(err.Error(), "scaffolder") {
+		t.Fatalf("expected scaffolder configuration error, got %v", err)
 	}
-	service := NewService(registry, nil, nil)
-
-	info, err := service.Info(context.Background(), "acme/sqlite")
-	if err != nil {
-		t.Fatal(err)
+	if _, err := service.Publish(ctx, modulepublish.Options{}); err == nil || !strings.Contains(err.Error(), "publisher") {
+		t.Fatalf("expected publisher configuration error, got %v", err)
 	}
-	if info.SelectedVersion != "1.0.0" || info.Newest != "2.0.0-rc.1" || info.Namespace != "DB::SQLITE" || info.Documentation != "https://registry.example/docs.md" {
-		t.Fatalf("unexpected stable info: %#v", info)
-	}
-
-	info, err = service.Info(context.Background(), "acme/preview")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if info.Latest != "" || info.SelectedVersion != "1.0.0-rc.1" || info.Newest != "1.0.0-rc.1" {
-		t.Fatalf("unexpected prerelease info: %#v", info)
-	}
-}
-
-func TestServiceInfoPropagatesBarnErrors(t *testing.T) {
-	service := NewService(&fakeRegistry{moduleErr: barnregistry.ErrModuleNotFound}, nil, nil)
-	if _, err := service.Info(context.Background(), "missing/module"); !errors.Is(err, barnregistry.ErrModuleNotFound) {
-		t.Fatalf("expected module-not-found error, got %v", err)
-	}
-
-	service = NewService(&fakeRegistry{modules: map[string]*barnregistry.Module{
-		"acme/empty": {ID: "acme/empty"},
-	}}, nil, nil)
-	if _, err := service.Info(context.Background(), "acme/empty"); !errors.Is(err, barnregistry.ErrMalformedArtifact) {
-		t.Fatalf("expected malformed-artifact error, got %v", err)
-	}
-}
-
-type fakeRegistry struct {
-	summaries     []barnregistry.ModuleSummary
-	modules       map[string]*barnregistry.Module
-	versions      map[string]*barnregistry.Version
-	search        func(context.Context, barnregistry.SearchOptions) ([]barnregistry.ModuleSummary, error)
-	searchOptions barnregistry.SearchOptions
-	searchErr     error
-	moduleErr     error
-	versionErr    error
-}
-
-func (registry *fakeRegistry) Search(ctx context.Context, options barnregistry.SearchOptions) ([]barnregistry.ModuleSummary, error) {
-	registry.searchOptions = options
-	if registry.search != nil {
-		return registry.search(ctx, options)
-	}
-
-	return registry.summaries, registry.searchErr
-}
-
-func (registry *fakeRegistry) Module(_ context.Context, id string) (*barnregistry.Module, error) {
-	if registry.moduleErr != nil {
-		return nil, registry.moduleErr
-	}
-
-	item, exists := registry.modules[id]
-	if !exists {
-		return nil, barnregistry.ErrModuleNotFound
-	}
-
-	return item, nil
-}
-
-func (registry *fakeRegistry) Version(_ context.Context, id, version string) (*barnregistry.Version, error) {
-	if registry.versionErr != nil {
-		return nil, registry.versionErr
-	}
-
-	item, exists := registry.versions[id+"@"+version]
-	if !exists {
-		return nil, barnregistry.ErrVersionNotFound
-	}
-
-	return item, nil
 }
